requirement.txt
python-dotenv
pandas
numpy
pymango
certifi
scikit-learn
